tests/wpt/web-platform-tests/webmessaging/with-options/null-transfer.html

<!doctype html>
<title>null transfer</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function(t) {
  assert_throws_js(TypeError, () => postMessage('', {transfer: null}));
});
</script>
